# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Quiz development",
    "description": "Quiz API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:9006",
  "basePath": "/",
  "paths": {
...
    "/api/district/addDistrict": {
      "post": {
        "summary": "Create a district",
        "tags": [
          "District"
        ],
        "description": "Create district",
        "parameters": [
          {
            "name": "Request Body",
            "in": "body",
            "required": true,
            "type": "object",
            "example": {
              "districtName": "Tenkasi",
              "isActive": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "District created successfully."
          }
        }
      }
    },
    "/api/district/updateDistrict/{id}": {
      "put": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "minimum": 1,
            "description": "ID of save district that we want to update",
            "type": "integer"
          },
          {
            "name": "Request Body",
            "in": "body",
            "required": true,
            "type": "object",
            "example": {
              "districtName" : "Tenkasi",
              "isActive" : true
            }
          }
        ],
        "tags": [
          "District"
        ],
        "summary": "Update district",
        "responses": {
          "200": {
            "description": "OK",
            "schema": ""
          }
        }
      }
    },
...
    "/api/partner/addPartner": {
      "post": {
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "description": "Bearer token"
          },
          {
            "name": "Request Body",
            "in": "body",
            "required": true,
            "description": "Add partner Details",
            "type": "object",
            "example": {
              "userID": 1,
              "partnerName": "Rambo",
              "partnerMobileNumber": 7890654321
            }
          }
        ],
        "tags": [
          "Partnar"
        ],
        "summary": "Add Partnar",
        "responses": {
          "200": {
            "description": "OK",
            "schema": ""
          }
        }
      }
    },
    "/api/partner/deletePartner/{id}": {
      "delete": {
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "description": "Bearer token"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "minimum": 1,
            "description": "ID of partnar that we want to delete",
            "type": "integer"
          }
        ],
        "tags": [
          "Partnar"
        ],
        "summary": "Delete Partnar",
        "responses": {
          "200": {
            "description": "OK",
            "schema": ""
          }
        }
      }
    },
    "/api/partner/getPartner": {
      "post": {
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "description": "Bearer token"
          },
          {
            "name": "Request Body",
            "in": "body",
            "required": true,
            "type": "object",
            "example": {
              "userID": 1
            }
          }
        ],
        "tags": [
          "Partnar"
        ],
        "summary": "Get Partner",
        "responses": {
          "200": {
            "description": "OK",
            "schema": ""
          }
        }
      }
    },
...
  },
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ]
}